refactor: extract reusble funcs for multi-adapters
